Restructure drivers
