+++ deno.json
@@ -1,5 +1,6 @@
 {
   "imports": {
-    "@/": "./src/"
+    "@/": "./src/",
+    "@db/sqlite": "jsr:@db/sqlite@^0.12.0"
   }
 }
