+++ package.json
@@ -49,6 +49,7 @@
     "sqlite3": "^5.1.6",
     "tsx": "^4.7.2",
     "typescript": "^5.4.5",
-    "uvu": "^0.5.6"
+    "uvu": "^0.5.6",
+    "@db/sqlite": "npm:@jsr/db__sqlite"
   }
 }
