+++ src/driver/jsr.ts
@@ -1,0 +1,1 @@
+export {connect as '@db/sqlite'} from './@db_sqlite.ts'
