+++ src/driver/npm.ts
@@ -1,0 +1,5 @@
+export {connect as '@electric-sql/pglite'} from './@electric-sql_pglite.ts'
+export {connect as 'better-sqlite3'} from './better-sqlite3.ts'
+export {connect as 'mysql2'} from './mysql2.ts'
+export {connect as 'pg'} from './pg.ts'
+export {connect as 'sql.js'} from './sql.js.ts'
