+++ test/driver/@electric-sql_pglite.test.ts
@@ -1,7 +1,7 @@
-import {connect} from '../../src/driver/pglite.ts'
 import {testDriver} from '../TestDriver.ts'
 
 await testDriver(import.meta, async () => {
+  const driver = await import('../../src/driver.ts')
   const {PGlite} = await import('@electric-sql/pglite')
-  return connect(new PGlite())
+  return driver['@electric-sql/pglite'](new PGlite())
 })
