+++ test/driver/better-sqlite3.test.ts
@@ -4,6 +4,6 @@ import {isNode} from '../TestRuntime.ts'
 if (isNode)
   await testDriver(import.meta, async () => {
     const {default: Database} = await import('better-sqlite3')
-    const {connect} = await import('../../src/driver/better-sqlite3.ts')
-    return connect(new Database(':memory:'))
+    const driver = await import('../../src/driver.ts')
+    return driver['better-sqlite3'](new Database(':memory:'))
   })
