+++ test/driver/bun-sqlite.test.ts
@@ -3,8 +3,8 @@ import {isBun} from '../TestRuntime.ts'
 
 async function createDb() {
   const {Database} = await import('bun:sqlite')
-  const {connect} = await import('../../src/driver/bun-sqlite.ts')
-  return connect(new Database(':memory:'))
+  const driver = await import('../../src/driver.ts')
+  return driver['bun:sqlite'](new Database(':memory:'))
 }
 
 if (isBun) await testDriver(import.meta, createDb)
