+++ test/driver/sql.js.test.ts
@@ -1,8 +1,8 @@
-import {connect} from '../../src/driver/sql.js.ts'
 import {testDriver} from '../TestDriver.ts'
 
 await testDriver(import.meta, async () => {
+  const driver = await import('../../src/driver.ts')
   const {default: init} = await import('sql.js')
   const {Database} = await init()
-  return connect(new Database())
+  return driver['sql.js'](new Database())
 })
